backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
ccxt>=4.0.0
pandas==2.1.3
numpy==1.26.2
python-telegram-bot==20.7
pydantic==2.5.0
python-dotenv==1.0.0
websockets==12.0
aiohttp==3.9.1
ta==0.11.0
